back/requirements.txt
Click==7.0
h11==0.8.1
httptools==0.0.13
python-engineio==3.9.3
python-socketio==4.3.1
six==1.12.0
uvicorn==0.9.0
uvloop==0.13.0
websockets==8.0.2
